Adds title and description for pages
